Database: added own Nette\DI\CompilerExtension for Database
